Move metrics class to its own file
